Update git ignore

+++ AndrickDLL/AndrickPlugin/AndrickPlugin/math/vector4.h
@@ -10,7 +10,7 @@ namespace ap
 
 
 	public:
-		Vector4(int x, int y, int z, int w) { this->x = x; this->y = y; this->z = z; this->w = w; };
+		Vector4(float x, float y, float z, float w) { this->x = x; this->y = y; this->z = z; this->w = w; };
 
 	};
 }
